Added more subclasses to increase specificity. Added Custom objects.
