+++ dx2d/Components/GraphicsComponent.cpp
@@ -1,0 +1,3 @@
+#include "GraphicsComponent.h"
+
+GraphicsComponent::GraphicsComponent(Type type) : Component(type) { }
