+++ dx2d/Components/GraphicsComponent.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../Graphics/GraphicsScene.h"
+#include "Component.h"
+
+class GraphicsComponent : public Component {
+public:
+	GraphicsComponent(Type type);
+};
