+++ dx2d/Components/LightComponent.cpp
@@ -1,0 +1,3 @@
+#include "LightComponent.h"
+
+LightComponent::LightComponent(Type type) : GraphicsComponent(type) { }
