+++ dx2d/Components/LightComponent.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "GraphicsComponent.h"
+
+class LightComponent : public GraphicsComponent {
+public:
+	LightComponent(Type type);
+};
