--- dx2d/Components/PhysicsBody.cpp
@@ -1,1 +1,0 @@
-#include "PhysicsBody.h"
