+++ dx2d/Objects/Engine/DefaultCube.cpp
@@ -1,0 +1,5 @@
+#include "DefaultCube.h"
+
+DefaultCube::DefaultCube() : Object() {
+	root_component.add_component(StaticMeshComponent());
+}
