+++ dx2d/Objects/Engine/DefaultCube.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../../Object.h"
+#include "../../Components/StaticMeshComponent.h"
+
+class DefaultCube : public Object {
+public:
+	DefaultCube();
+};
+
