+++ dx2d/dx2d.vcxproj
@@ -156,16 +156,17 @@
     <ClInclude Include="Components\Component.h" />
     <ClInclude Include="Components\DirectionalLightComponent.h" />
     <ClInclude Include="Components\PhysicalComponent.h" />
-    <ClInclude Include="Components\PhysicsBody.h" />
+    <ClInclude Include="PhysicsBody.h" />
     <ClInclude Include="Components\PointLightComponent.h" />
     <ClInclude Include="Components\SpotlightComponent.h" />
     <ClInclude Include="Components\StaticMeshComponent.h" />
+    <ClInclude Include="Objects\Engine\DefaultCube.h" />
     <ClInclude Include="Engine.h" />
     <ClInclude Include="globalstructs.h" />
     <ClInclude Include="globalutil.h" />
     <ClInclude Include="Graphics\d3dx12.h" />
     <ClInclude Include="Graphics\GraphicsPipeline.h" />
-    <ClInclude Include="Graphics\GraphicsPipelineMeshProxy.h" />
+    <ClInclude Include="Graphics\GraphicsPipelineMeshChangeManager.h" />
     <ClInclude Include="Graphics\GraphicsScene.h" />
     <ClInclude Include="Graphics\Material.h" />
     <ClInclude Include="Graphics\Texture.h" />
@@ -183,13 +184,14 @@
     <ClCompile Include="Components\Component.cpp" />
     <ClCompile Include="Components\DirectionalLightComponent.cpp" />
     <ClCompile Include="Components\PhysicalComponent.cpp" />
-    <ClCompile Include="Components\PhysicsBody.cpp" />
+    <ClCompile Include="PhysicsBody.cpp" />
     <ClCompile Include="Components\PointLightComponent.cpp" />
     <ClCompile Include="Components\SpotlightComponent.cpp" />
     <ClCompile Include="Components\StaticMeshComponent.cpp" />
+    <ClCompile Include="Objects\Engine\DefaultCube.cpp" />
     <ClCompile Include="Engine.cpp" />
     <ClCompile Include="Graphics\GraphicsPipeline.cpp" />
-    <ClCompile Include="Graphics\GraphicsPipelineMeshProxy.cpp" />
+    <ClCompile Include="Graphics\GraphicsPipelineMeshChangeManager.cpp" />
     <ClCompile Include="Graphics\GraphicsScene.cpp" />
     <ClCompile Include="Graphics\Material.cpp" />
     <ClCompile Include="Graphics\Texture.cpp" />
